+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## [Version - 0]
 
+## improve [v0.39.13] 2023-11-8
+
+- call api add/edit special point menu
+
 ## Fix [v0.39.12] 2023-11-08 
 - fixed payload pin news 
+
 ## Fix [v0.39.11] 2023-11-6
 
 - fixed target spray in all task menu